# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de prueba",
        "description": "Api de prueba",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/tareas": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Listado de usuarios",
                "operationId": "2493988f69d01796099b082aa85966f7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "id_usuario": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "prueba"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "prueba"
                                                    },
                                                    "fecha_limite": {
                                                        "type": "string",
                                                        "example": "2023-02-23"
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "Pendiente"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "7b25afab68f5e724c98dc510c5df2217",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titulo": {
                                        "type": "string",
                                        "example": "Tarea 1"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripcion de la tarea"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Pendiente"
                                    }
                                },
                                "type": "object"
                            },
                            "example": "2023-02-23"
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    },
                    "400": {
                        "description": "BAD REQUEST"
                    }
                }
            }
        },
        "/api/tareas/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Mostrar una tarea específica",
                "operationId": "33adaac00a557cf973e21ad403ca8ba5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_usuario": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "prueba"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "prueba"
                                        },
                                        "fecha_limite": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-6-7"
                                        },
                                        "estado": {
                                            "type": "string",
                                            "example": "pendiente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea específica",
                "operationId": "6f8e8a5de70a166605458b48edf1feed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer"
                                    },
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "fecha_limite": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "estado": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "BAD REQUEST"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea específica",
                "operationId": "5c3c6638cc4411907241c6ff71f2b971",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "BAD REQUEST"
                    }
                }
            }
        },
...
    }
}